refactor(ui): Extract the create run form from its route
Move form state, validation, field rendering, and payload conversion into
an exported `CreateRunForm` component with explicit input and submission
props. This gives the form a bounded interface that can be tested without
mounting the route.

Keep loader data, queries, mutation side effects, toasts, cache updates,
and navigation in `CreateRunPage`. The route now passes loaded values into
the form and submits the resulting `PostRepositoryRun` payload.
